+++ scripts/upgradeSlave.sh
@@ -22,9 +22,10 @@ else
   exit 1
 fi
 
+BRANCH="master"
 DL_URL="https://github.com/MrKaKisen/nafsdm/archive/"
 MY_VERSION_RAW="`cat /home/slave-nafsdm/pythondaemon/version.py`"
-LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/master/version.txt)
+LATEST_VERSION=$(curl https://raw.githubusercontent.com/MrKaKisen/nafsdm/$BRANCH/version.txt)
 
 # determine supported versions
 if [ "$MY_VERSION_RAW" == 'version = "1.0.1-stable"' ]; then
@@ -39,6 +40,9 @@ elif [ "$MY_VERSION_RAW" == 'version = "1.2-stable"' ]; then
 elif [ "$MY_VERSION_RAW" == 'version = "1.2.1-stable"' ]; then
   echo "* Detected version 1.2.1-stable - supported by this upgrade script."
   MY_VERSION="1.2.1-stable"
+elif [ "$MY_VERSION_RAW" == 'version = "1.2.2-stable"' ]; then
+  echo "* Detected version 1.2.2-stable - supported by this upgrade script."
+  MY_VERSION="1.2.2-stable"
 else
   echo "* Your version is not supported (dev versions and 1.0 is not supported)."
   exit 1
@@ -120,6 +124,14 @@ elif [ "$MY_VERSION" == "1.2.1-stable" ]; then
   rm -rf /home/slave-nafsdm/start.py
   cp nafsdm/slave-nafsdm/start.py /home/slave-nafsdm/start.py -R
   chmod +x /home/slave-nafsdm/start.py
+
+  echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
+  rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh
+
+elif [ "$MY_VERSION" == "1.2.2-stable" ]; then
+  echo "* Replacing python files.."
+  rm -rf /home/slave-nafsdm/pythondaemon
+  cp nafsdm/slave-nafsdm/pythondaemon /home/slave-nafsdm/pythondaemon -R
 
   echo "* Upgrade completed. You can now start nafsdm-slave again (make sure master is also upgraded!)-"
   rm -rf /home/slave-nafsdm/pythondaemon/tempUpgrade/temp_upgrade.sh