+++ slave-nafsdm/pythondaemon/__main__.py
@@ -5,6 +5,8 @@
 
 import logging
 import sys
+import shutil
+import os
 from daemon import *
 from version import version
 from getConfig import getConfig
@@ -43,6 +45,16 @@ def main():
     if os.path.isfile("/home/slave-nafsdm/config-legacy.conf"):
         logging.warning("Legacy config was found. This probably means that nafsdm has been upgraded but the config hasn't been changed yet.")
         exit(1)
+
+    # check if the temp folder exists
+    if os.path.isdir("/home/slave-nafsdm/temp"):
+        logging.debug("Temp folder already exists.")
+    else:
+        if os.path.isfile("/home/slave-nafsdm/temp"):
+            os.remove("/home/slave-nafsdm/temp")
+            os.makedirs("/home/slave-nafsdm/temp")
+        else:
+            os.makedirs("/home/slave-nafsdm/temp")
 
     # get config
     config = getConfig()