+++ slave-nafsdm/pythondaemon/daemon.py
@@ -8,88 +8,119 @@ import time
 import os
 import sys
 import subprocess
+from db import parseDbData
+from shutil import copyfile
+
+def changeDetected():
+    changeDetected = None
+    beforeExists = None
+    tempExists = None
+
+    if os.path.isfile("/home/slave-nafsdm/temp/domains_before.sql"):
+        f = open("/home/slave-nafsdm/temp/domains_before.sql")
+        domainsBefore = f.read()
+        f.close()
+        beforeExists = True
+    else:
+        logging.warning("Couldn't read from before domains temp file. Is this the first time running maybe?")
+        beforeExists = False
+
+    if os.path.isfile("/home/slave-nafsdm/temp/domains_temp.sql"):
+        f = open("/home/slave-nafsdm/temp/domains_temp.sql")
+        domainsNew = f.read()
+        f.close()
+        tempExists = True
+    else:
+        logging.warning("Couldn't read from domains temp file. Is this the first time running maybe?")
+        tempExists = False
+
+    if tempExists == True and beforeExists == True:
+        if domainsBefore != domainsNew:
+            changeDetected = True
+        else:
+            changeDetected = False
+    else:
+        changeDetected = False
+
+    return changeDetected
 
 def getData(config):
     try:
-         outputNull = subprocess.check_output(["scp", "-i", "/home/slave-nafsdm/.ssh/master_key", config.user + "@" + config.host + ":/home/master-nafsdm/data/domains.txt", "/home/slave-nafsdm/domains.temp"])
+         outputNull = subprocess.check_output(["scp", "-i", "/home/slave-nafsdm/.ssh/master_key", config.user + "@" + config.host + ":/home/master-nafsdm/data/domains.sql", "/home/slave-nafsdm/temp/domains_temp.sql"])
     except Exception:
         logging.exception("An error occured during SCP connection.")
         logging.error("Please check if the master is up and working.")
 
-# find slave function
-def find_slave(currentLine, myhostname):
-    if not len(currentLine.split()) < 2:
-        # split it by space
-        split1 = currentLine.split()
-
-        wasFound = False
-        for currentSlave in split1[3].split("."):
-            if currentSlave == myhostname:
-                wasFound = True
-
-        return wasFound
-
 def writeData(config):
-    if os.path.isfile("/home/slave-nafsdm/domains.temp") == True:
-        f = open("/home/slave-nafsdm/domains.temp")
-        domainsData = f.read()
-        f.close()
+    if os.path.isfile("/home/slave-nafsdm/temp/domains_temp.sql") == True:
+        domainsToWrite = parseDbData(config)
+        if domainsToWrite != False:
 
-        # remove config temporarily
-        if os.path.isfile(config.bindPath):
-            os.remove(config.bindPath)
+            # remove config temporarily
+            if os.path.isfile(config.bindPath):
+                logging.debug("Removing config temporarily")
+                os.remove(config.bindPath)
 
-        for currentLine in domainsData.split("\n"):
-            if len(currentLine.split()) == 5:
-                wasFound = find_slave(currentLine, config.nodeName)
-                if (wasFound == True):
-                    if len(currentLine.split()[4]) != 2:
-                        if currentLine.split()[4].split(".")[1] == "yes":
-                            f = open(config.bindPath, "a")
-                            if config.type == "debian" or config.type == "ubuntu":
-                                f.write('''/* ''' + currentLine.split()[2] + ''' */
-zone "''' + currentLine.split()[0] + '''" IN {
+            # test if there is a folder available
+            try:
+                ftest = open(config.bindPath, "w")
+            except Exception:
+                logging.exception("Couldn't write to bind file.")
+                logging.error("Please check if the folder exists.")
+
+                return False
+
+            invalidSystemType = False
+            for r in domainsToWrite:
+                if r[5] == "y":
+                    f = open(config.bindPath, "a")
+                    if config.type == "debian" or config.type == "ubuntu":
+                        f.write('''/* ''' + r[3] + ''' */
+zone "''' + r[1] + '''" IN {
     type slave;
-    file "db.''' + currentLine.split()[0] + '''.signed";
-    masters { ''' + currentLine.split()[1] + '''; };
+    file "db.''' + r[1] + '''.signed";
+    masters { ''' + r[2] + '''; };
 }; ''' + "\n" + "\n")
-                                f.close()
-                            elif config.type == "centos":
-                                f.write('''/* ''' + currentLine.split()[2] + ''' */
-zone "''' + currentLine.split()[0] + '''" IN {
+                        f.close()
+                    elif config.type == "centos":
+                        f.write('''/* ''' + r[3] + ''' */
+zone "''' + r[1] + '''" IN {
     type slave;
-    file "slaves/''' + currentLine.split()[0] + '''.signed";
-    masters { ''' + currentLine.split()[1] + '''; };
+    file "slaves/''' + r[1] + '''.signed";
+    masters { ''' + r[2] + '''; };
 }; ''' + "\n" + "\n")
-                                f.close()
-                        elif currentLine.split()[4].split(".")[1] == "no":
-                            f = open(config.bindPath, "a")
-                            if config.type == "debian" or config.type == "ubuntu":
-                                f.write('''/* ''' + currentLine.split()[2] + ''' */
-zone "''' + currentLine.split()[0] + '''" IN {
-    type slave;
-    file "db.''' + currentLine.split()[0] + '''";
-    masters { ''' + currentLine.split()[1] + '''; };
-}; ''' + "\n" + "\n")
-                                f.close()
-                            elif config.type == "centos":
-                                f.write('''/* ''' + currentLine.split()[2] + ''' */
-zone "''' + currentLine.split()[0] + '''" IN {
-    type slave;
-    file "slaves/''' + currentLine.split()[0] + '''";
-    masters { ''' + currentLine.split()[1] + '''; };
-}; ''' + "\n" + "\n")
-                                f.close()
+                        f.close()
                     else:
-                        logging.critical("Invalid system type. Debian (ubuntu) & CentOS only supported.")
-                        exit(1)
+                        invalidSystemType = True
+                elif r[5] == "n":
+                    f = open(config.bindPath, "a")
+                    if config.type == "debian" or config.type == "ubuntu":
+                        f.write('''/* ''' + r[3] + ''' */
+zone "''' + r[1] + '''" IN {
+    type slave;
+    file "db.''' + r[1] + '''";
+    masters { ''' + r[2] + '''; };
+}; ''' + "\n" + "\n")
+                        f.close()
+                    elif config.type == "centos":
+                        f.write('''/* ''' + r[3] + ''' */
+zone "''' + r[1] + '''" IN {
+    type slave;
+    file "slaves/''' + r[1] + '''";
+    masters { ''' + r[2] + '''; };
+}; ''' + "\n" + "\n")
+                        f.close()
+                    else:
+                        invalidSystemType = True
+            if invalidSystemType == True:
+                logging.critical("Invalid system type! Please check your config!")
+                exit(1)
+            logging.debug("New config has been written.")
     else:
         logging.error("An error occured while reading data that was recently downloaded. This usually means the file never was downloaded and therefore doesn't exist.")
         logging.error("Couldn't read from domains file! Connection error?")
 
-def commandReload(domainsNew):
-    # just to split things up
-
+def reloadBind():
     # if it fails, it will be printed in log
     reloadSucceeded = True
     try:
@@ -101,40 +132,11 @@ def commandReload(domainsNew):
 
     if reloadSucceeded == True:
         # update the before file so reload doesn't occur again
-        f = open("/home/slave-nafsdm/domains.before", "w")
-        f.write(domainsNew)
-        f.close()
-
-def reloadBind():
-    continueReload = True
-    beforeExists = True
-    if os.path.isfile("/home/slave-nafsdm/domains.before"):
-        f = open("/home/slave-nafsdm/domains.before")
-        domainsBefore = f.read()
-        f.close()
-    else:
-        logging.warning("Couldn't read from before domains temp file. Is this the first time running maybe?")
-        continueReload = True
-        beforeExists = False
-
-    if os.path.isfile("/home/slave-nafsdm/domains.temp"):
-        f = open("/home/slave-nafsdm/domains.temp")
-        domainsNew = f.read()
-        f.close()
-    else:
-        logging.warning("Couldn't read from domains temp file. Is this the first time running maybe?")
-        continueReload = False
-
-    if (continueReload == True):
-        if (beforeExists == True):
-            if domainsBefore != domainsNew:
-                logging.info("Change detected! Reloading bind..")
-                commandReload(domainsNew)
-        else:
-            logging.info("No before file. Reloading bind..")
-            commandReload(domainsNew)
-    else:
-        logging.error("Bind reload aborted due to earlier errors.")
+        try:
+            copyfile("/home/slave-nafsdm/temp/domains_temp.sql", "/home/slave-nafsdm/temp/domains_before.sql")
+        except Exception:
+            logging.exception("Error occured during file replacement domains_temp to domains_before.")
+            logging.error("Do we have permissions to that folder?")
 
 
 def runDaemon(config):
@@ -150,5 +152,8 @@ def runDaemon(config):
         time.sleep(int(config.update_interval))
 
         getData(config)
-        writeData(config)
-        reloadBind()
+        changeStatus = changeDetected()
+        if changeStatus == True:
+            logging.info("Change detected! Writing configuration & reloading bind")
+            writeData(config)
+            reloadBind()